requirements.txt
mkdocs
mkdocs-material
mdx_truly_sane_lists
mkdocs-video
markdown-exec[ansi]
Flask~=3.0.2
pyyaml